+++ src/carousel/Carousel.js
@@ -1,13 +1,23 @@
 import React, { Component } from 'react';
-import { View, FlatList, Animated, Platform, I18nManager, ViewPropTypes } from 'react-native';
+import { Animated, Easing, FlatList, I18nManager, Platform, ScrollView, View, ViewPropTypes } from 'react-native';
 import PropTypes from 'prop-types';
 import shallowCompare from 'react-addons-shallow-compare';
+import {
+    defaultScrollInterpolator,
+    stackScrollInterpolator,
+    tinderScrollInterpolator,
+    defaultAnimatedStyles,
+    shiftAnimatedStyles,
+    stackAnimatedStyles,
+    tinderAnimatedStyles
+} from '../utils/animations';
 
 const IS_IOS = Platform.OS === 'ios';
 
 // Native driver for scroll events
 // See: https://facebook.github.io/react-native/blog/2017/02/14/using-native-driver-for-animated.html
-const AnimatedFlatList = Animated.createAnimatedComponent(FlatList);
+const AnimatedFlatList = FlatList ? Animated.createAnimatedComponent(FlatList) : null;
+const AnimatedScrollView = Animated.createAnimatedComponent(ScrollView);
 
 // React Native automatically handles RTL layouts; unfortunately, it's buggy with horizontal ScrollView
 // See https://github.com/facebook/react-native/issues/11960
@@ -18,13 +28,14 @@ const IS_RTL = I18nManager.isRTL;
 export default class Carousel extends Component {
 
     static propTypes = {
-        ...FlatList.propTypes,
         data: PropTypes.array.isRequired,
         renderItem: PropTypes.func.isRequired,
         itemWidth: PropTypes.number, // required for horizontal carousel
         itemHeight: PropTypes.number, // required for vertical carousel
-        sliderWidth: PropTypes.number,  // required for horizontal carousel
+        sliderWidth: PropTypes.number, // required for horizontal carousel
         sliderHeight: PropTypes.number, // required for vertical carousel
+        activeAnimationType: PropTypes.string,
+        activeAnimationOptions: PropTypes.object,
         activeSlideAlignment: PropTypes.oneOf(['center', 'end', 'start']),
         activeSlideOffset: PropTypes.number,
         apparitionDelay: PropTypes.number,
@@ -41,22 +52,32 @@ export default class Carousel extends Component {
         inactiveSlideOpacity: PropTypes.number,
         inactiveSlideScale: PropTypes.number,
         inactiveSlideShift: PropTypes.number,
+        layout: PropTypes.oneOf(['default', 'stack', 'tinder']),
+        layoutCardOffset: PropTypes.number,
+        lockScrollTimeoutDuration: PropTypes.number,
         lockScrollWhileSnapping: PropTypes.bool,
         loop: PropTypes.bool,
         loopClonesPerSide: PropTypes.number,
-        slideStyle: Animated.View.propTypes.style,
+        scrollEnabled: PropTypes.bool,
+        scrollInterpolator: PropTypes.func,
+        slideInterpolatedStyle: PropTypes.func,
+        slideStyle: ViewPropTypes ? ViewPropTypes.style : View.propTypes.style,
         shouldOptimizeUpdates: PropTypes.bool,
         swipeThreshold: PropTypes.number,
+        useScrollView: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
         vertical: PropTypes.bool,
+        onBeforeSnapToItem: PropTypes.func,
         onSnapToItem: PropTypes.func
     };
 
     static defaultProps = {
+        activeAnimationType: 'timing',
+        activeAnimationOptions: null,
         activeSlideAlignment: 'center',
         activeSlideOffset: 20,
-        apparitionDelay: 250,
+        apparitionDelay: 0,
         autoplay: false,
-        autoplayDelay: 5000,
+        autoplayDelay: 1000,
         autoplayInterval: 3000,
         callbackOffsetMargin: 5,
         containerCustomStyle: {},
@@ -68,12 +89,16 @@ export default class Carousel extends Component {
         inactiveSlideOpacity: 0.7,
         inactiveSlideScale: 0.9,
         inactiveSlideShift: 0,
+        layout: 'default',
+        lockScrollTimeoutDuration: 1000,
         lockScrollWhileSnapping: false,
         loop: false,
         loopClonesPerSide: 3,
+        scrollEnabled: true,
         slideStyle: {},
         shouldOptimizeUpdates: true,
         swipeThreshold: 20,
+        useScrollView: !AnimatedFlatList,
         vertical: false
     }
 
@@ -93,12 +118,15 @@ export default class Carousel extends Component {
         this._previousFirstItem = initialActiveItem;
         this._previousItemsLength = initialActiveItem;
 
+        this._mounted = false;
         this._positions = [];
         this._currentContentOffset = 0; // store ScrollView's scroll position
+        this._canFireBeforeCallback = false;
         this._canFireCallback = false;
         this._scrollOffsetRef = null;
         this._onScrollTriggered = true; // used when momentum is enabled to prevent an issue with edges items
-        this._scrollEnabled = props.scrollEnabled === false ? false : true;
+        this._lastScrollDate = 0; // used to work around a FlatList bug
+        this._scrollEnabled = props.scrollEnabled !== false;
 
         this._initPositionsAndInterpolators = this._initPositionsAndInterpolators.bind(this);
         this._renderItem = this._renderItem.bind(this);
@@ -111,22 +139,12 @@ export default class Carousel extends Component {
         this._onScrollEndDrag = !props.enableMomentum ? this._onScrollEndDrag.bind(this) : undefined;
         this._onMomentumScrollEnd = props.enableMomentum ? this._onMomentumScrollEnd.bind(this) : undefined;
         this._onTouchStart = this._onTouchStart.bind(this);
+        this._onTouchEnd = this._onTouchEnd.bind(this);
         this._onTouchRelease = this._onTouchRelease.bind(this);
 
-        // Native driver for scroll events
-        const scrollEventConfig = {
-            listener: this._onScroll,
-            useNativeDriver: true
-        };
-        this._scrollPos = new Animated.Value(0);
-        this._onScrollHandler = props.vertical ?
-            Animated.event(
-                [{ nativeEvent: { contentOffset: { y: this._scrollPos } } }],
-                scrollEventConfig
-            ) : Animated.event(
-                [{ nativeEvent: { contentOffset: { x: this._scrollPos } } }],
-                scrollEventConfig
-            );
+        this._getKeyExtractor = this._getKeyExtractor.bind(this);
+
+        this._setScrollHandler(props);
 
         // This bool aims at fixing an iOS bug due to scrollTo that triggers onMomentumScrollEnd.
         // onMomentumScrollEnd fires this._snapScroll, thus creating an infinite loop.
@@ -137,29 +155,52 @@ export default class Carousel extends Component {
             console.warn('react-native-snap-carousel: It is recommended to use at least version 0.44 of React Native with the plugin');
         }
         if (!props.vertical && (!props.sliderWidth || !props.itemWidth)) {
-            console.warn('react-native-snap-carousel: You need to specify both `sliderWidth` and `itemWidth` for horizontal carousels');
+            console.error('react-native-snap-carousel: You need to specify both `sliderWidth` and `itemWidth` for horizontal carousels');
         }
         if (props.vertical && (!props.sliderHeight || !props.itemHeight)) {
-            console.warn('react-native-snap-carousel: You need to specify both `sliderHeight` and `itemHeight` for vertical carousels');
+            console.error('react-native-snap-carousel: You need to specify both `sliderHeight` and `itemHeight` for vertical carousels');
+        }
+        if (props.apparitionDelay && !IS_IOS && !props.useScrollView) {
+            console.warn('react-native-snap-carousel: Using `apparitionDelay` on Android is not recommended since it can lead to rendering issues');
+        }
+        if (props.customAnimationType || props.customAnimationOptions) {
+            console.warn('react-native-snap-carousel: Props `customAnimationType` and `customAnimationOptions` have been renamed to `activeAnimationType` and `activeAnimationOptions`');
         }
         if (props.onScrollViewScroll) {
-            console.warn('react-native-snap-carousel: Prop `onScrollViewScroll` has been removed. Use `onScroll` instead');
+            console.error('react-native-snap-carousel: Prop `onScrollViewScroll` has been removed. Use `onScroll` instead');
         }
     }
 
     componentDidMount () {
-        const { apparitionDelay } = this.props;
+        const { apparitionDelay, autoplay, firstItem } = this.props;
+        const _firstItem = this._getFirstItem(firstItem);
+        const apparitionCallback = () => {
+            this.setState({ hideCarousel: false });
+            if (autoplay) {
+                this.startAutoplay();
+            }
+        };
 
+        this._mounted = true;
         this._initPositionsAndInterpolators();
 
-        if (apparitionDelay) {
-            // Hide FlatList's awful init
-            this._apparitionTimeout = setTimeout(() => {
-                this._didMountDelayedInit();
-            }, apparitionDelay);
-        } else {
-            this._didMountDelayedInit();
-        }
+        // Without 'requestAnimationFrame' or a `0` timeout, images will randomly not be rendered on Android...
+        requestAnimationFrame(() => {
+            if (!this._mounted) {
+                return;
+            }
+
+            this._snapToItem(_firstItem, false, false, true, false);
+            this._hackActiveSlideAnimation(_firstItem, 'start', true);
+
+            if (apparitionDelay) {
+                this._apparitionTimeout = setTimeout(() => {
+                    apparitionCallback();
+                }, apparitionDelay);
+            } else {
+                apparitionCallback();
+            }
+        });
     }
 
     shouldComponentUpdate (nextProps, nextState) {
@@ -170,26 +211,32 @@ export default class Carousel extends Component {
         }
     }
 
-    componentWillReceiveProps (nextProps) {
+    componentDidUpdate (prevProps) {
         const { interpolators } = this.state;
-        const { firstItem, itemHeight, itemWidth, sliderHeight, sliderWidth } = nextProps;
-        const itemsLength = this._getCustomDataLength(nextProps);
+        const { firstItem, itemHeight, itemWidth, scrollEnabled, sliderHeight, sliderWidth } = this.props;
+        const itemsLength = this._getCustomDataLength(this.props);
 
         if (!itemsLength) {
             return;
         }
 
-        const nextFirstItem = this._getFirstItem(firstItem, nextProps);
+        const nextFirstItem = this._getFirstItem(firstItem, this.props);
         let nextActiveItem = this._activeItem || this._activeItem === 0 ? this._activeItem : nextFirstItem;
 
-        const hasNewSliderWidth = sliderWidth && sliderWidth !== this.props.sliderWidth;
-        const hasNewSliderHeight = sliderHeight && sliderHeight !== this.props.sliderHeight;
-        const hasNewItemWidth = itemWidth && itemWidth !== this.props.itemWidth;
-        const hasNewItemHeight = itemHeight && itemHeight !== this.props.itemHeight;
+        const hasNewSliderWidth = sliderWidth && sliderWidth !== prevProps.sliderWidth;
+        const hasNewSliderHeight = sliderHeight && sliderHeight !== prevProps.sliderHeight;
+        const hasNewItemWidth = itemWidth && itemWidth !== prevProps.itemWidth;
+        const hasNewItemHeight = itemHeight && itemHeight !== prevProps.itemHeight;
+        const hasNewScrollEnabled = scrollEnabled !== prevProps.scrollEnabled;
 
         // Prevent issues with dynamically removed items
         if (nextActiveItem > itemsLength - 1) {
             nextActiveItem = itemsLength - 1;
+        }
+
+        // Handle changing scrollEnabled independent of user -> carousel interaction
+        if (hasNewScrollEnabled) {
+            this._setScrollEnabled(scrollEnabled);
         }
 
         if (interpolators.length !== itemsLength || hasNewSliderWidth ||
@@ -197,7 +244,7 @@ export default class Carousel extends Component {
             this._activeItem = nextActiveItem;
             this._previousItemsLength = itemsLength;
 
-            this._initPositionsAndInterpolators(nextProps);
+            this._initPositionsAndInterpolators(this.props);
 
             // Handle scroll issue when dynamically removing items (see #133)
             // This also fixes first item's active state on Android
@@ -212,11 +259,16 @@ export default class Carousel extends Component {
         } else if (nextFirstItem !== this._previousFirstItem && nextFirstItem !== this._activeItem) {
             this._activeItem = nextFirstItem;
             this._previousFirstItem = nextFirstItem;
-            this._snapToItem(nextFirstItem, true, true, false, false);
+            this._snapToItem(nextFirstItem, false, true, false, false);
+        }
+
+        if (this.props.onScroll !== prevProps.onScroll) {
+          this._setScrollHandler(this.props);
         }
     }
 
     componentWillUnmount () {
+        this._mounted = false;
         this.stopAutoplay();
         clearTimeout(this._apparitionTimeout);
         clearTimeout(this._hackSlideAnimationTimeout);
@@ -239,9 +291,39 @@ export default class Carousel extends Component {
         return this._currentContentOffset;
     }
 
-    _shouldAnimateSlides (props = this.props) {
-        const { inactiveSlideOpacity, inactiveSlideScale, inactiveSlideShift } = props;
-        return inactiveSlideOpacity < 1 || inactiveSlideScale < 1 || inactiveSlideShift !== 0;
+    _setScrollHandler(props) {
+      // Native driver for scroll events
+      const scrollEventConfig = {
+        listener: this._onScroll,
+        useNativeDriver: true,
+      };
+      this._scrollPos = new Animated.Value(0);
+      const argMapping = props.vertical
+        ? [{ nativeEvent: { contentOffset: { y: this._scrollPos } } }]
+        : [{ nativeEvent: { contentOffset: { x: this._scrollPos } } }];
+
+      if (props.onScroll && Array.isArray(props.onScroll._argMapping)) {
+        // Because of a react-native issue https://github.com/facebook/react-native/issues/13294
+        argMapping.pop();
+        const [ argMap ] = props.onScroll._argMapping;
+        if (argMap && argMap.nativeEvent && argMap.nativeEvent.contentOffset) {
+          // Shares the same animated value passed in props
+          this._scrollPos =
+            argMap.nativeEvent.contentOffset.x ||
+            argMap.nativeEvent.contentOffset.y ||
+            this._scrollPos;
+        }
+        argMapping.push(...props.onScroll._argMapping);
+      }
+      this._onScrollHandler = Animated.event(
+        argMapping,
+        scrollEventConfig
+      );
+    }
+
+    _needsScrollView () {
+        const { useScrollView } = this.props;
+        return useScrollView || !AnimatedFlatList || this._shouldUseStackLayout() || this._shouldUseTinderLayout();
     }
 
     _needsRTLAdaptations () {
@@ -250,20 +332,49 @@ export default class Carousel extends Component {
     }
 
     _canLockScroll () {
-        const { enableMomentum, lockScrollWhileSnapping } = this.props;
-        return !enableMomentum && lockScrollWhileSnapping;
+        const { scrollEnabled, enableMomentum, lockScrollWhileSnapping } = this.props;
+        return scrollEnabled && !enableMomentum && lockScrollWhileSnapping;
     }
 
     _enableLoop () {
         const { data, enableSnap, loop } = this.props;
-        return enableSnap && loop && data.length && data.length > 1;
+        return enableSnap && loop && data && data.length && data.length > 1;
+    }
+
+    _shouldAnimateSlides (props = this.props) {
+        const { inactiveSlideOpacity, inactiveSlideScale, scrollInterpolator, slideInterpolatedStyle } = props;
+        return inactiveSlideOpacity < 1 ||
+            inactiveSlideScale < 1 ||
+            !!scrollInterpolator ||
+            !!slideInterpolatedStyle ||
+            this._shouldUseShiftLayout() ||
+            this._shouldUseStackLayout() ||
+            this._shouldUseTinderLayout();
+    }
+
+    _shouldUseCustomAnimation () {
+        const { activeAnimationOptions } = this.props;
+        return !!activeAnimationOptions && !this._shouldUseStackLayout() && !this._shouldUseTinderLayout();
+    }
+
+    _shouldUseShiftLayout () {
+        const { inactiveSlideShift, layout } = this.props;
+        return layout === 'default' && inactiveSlideShift !== 0;
+    }
+
+    _shouldUseStackLayout () {
+        return this.props.layout === 'stack';
+    }
+
+    _shouldUseTinderLayout () {
+        return this.props.layout === 'tinder';
     }
 
     _getCustomData (props = this.props) {
         const { data, loopClonesPerSide } = props;
-        const dataLength = data.length;
+        const dataLength = data && data.length;
 
-        if (!data || !dataLength) {
+        if (!dataLength) {
             return [];
         }
 
@@ -369,29 +480,37 @@ export default class Carousel extends Component {
         return this._enableLoop() ? index + loopClonesPerSide : index;
     }
 
+    _getWrappedRef () {
+        // https://github.com/facebook/react-native/issues/10635
+        // https://stackoverflow.com/a/48786374/8412141
+        return this._carouselRef && this._carouselRef.getNode && this._carouselRef.getNode();
+    }
+
     _getScrollEnabled () {
         return this._scrollEnabled;
     }
 
-    _setScrollEnabled (value = true) {
-        if (this.props.scrollEnabled === false || !this._flatlist || !this._flatlist.setNativeProps) {
+    _setScrollEnabled (scrollEnabled = true) {
+        const wrappedRef = this._getWrappedRef();
+
+        if (!wrappedRef || !wrappedRef.setNativeProps) {
             return;
         }
 
         // 'setNativeProps()' is used instead of 'setState()' because the latter
         // really takes a toll on Android behavior when momentum is disabled
-        this._flatlist.setNativeProps({ scrollEnabled: value });
-        this._scrollEnabled = value;
+        wrappedRef.setNativeProps({ scrollEnabled });
+        this._scrollEnabled = scrollEnabled;
     }
 
     _getKeyExtractor (item, index) {
-        return `carousel-item-${index}`;
+        return this._needsScrollView() ? `scrollview-item-${index}` : `flatlist-item-${index}`;
     }
 
     _getScrollOffset (event) {
         const { vertical } = this.props;
         return (event && event.nativeEvent && event.nativeEvent.contentOffset &&
-            Math.round(event.nativeEvent.contentOffset[vertical ? 'y' : 'x'])) || 0;
+            event.nativeEvent.contentOffset[vertical ? 'y' : 'x']) || 0;
     }
 
     _getContainerInnerMargin (opposite = false) {
@@ -408,7 +527,7 @@ export default class Carousel extends Component {
         }
     }
 
-    _getViewportOffet () {
+    _getViewportOffset () {
         const { sliderWidth, sliderHeight, itemWidth, itemHeight, vertical, activeSlideAlignment } = this.props;
 
         if (activeSlideAlignment === 'start') {
@@ -423,7 +542,7 @@ export default class Carousel extends Component {
     }
 
     _getCenter (offset) {
-        return offset + this._getViewportOffet() - this._getContainerInnerMargin();
+        return offset + this._getViewportOffset() - this._getContainerInnerMargin();
     }
 
     _getActiveItem (offset) {
@@ -446,24 +565,11 @@ export default class Carousel extends Component {
         return 0;
     }
 
-    _didMountDelayedInit () {
-        const { firstItem, autoplay } = this.props;
-        const _firstItem = this._getFirstItem(firstItem);
-
-        this._snapToItem(_firstItem, false, false, true, false);
-        this._hackActiveSlideAnimation(_firstItem, 'start', true);
-        this.setState({ hideCarousel: false });
-
-        if (autoplay) {
-            this.startAutoplay();
-        }
-    }
-
     _initPositionsAndInterpolators (props = this.props) {
-        const { data, itemWidth, itemHeight, vertical } = props;
+        const { data, itemWidth, itemHeight, scrollInterpolator, vertical } = props;
         const sizeRef = vertical ? itemHeight : itemWidth;
 
-        if (!data.length) {
+        if (!data || !data.length) {
             return;
         }
 
@@ -472,19 +578,37 @@ export default class Carousel extends Component {
 
         this._getCustomData(props).forEach((itemData, index) => {
             const _index = this._getCustomIndex(index, props);
-            const start = (_index - 1) * sizeRef;
-            const middle = _index * sizeRef;
-            const end = (_index + 1) * sizeRef;
-            const animatedValue = this._shouldAnimateSlides(props) ? this._scrollPos.interpolate({
-                inputRange: [start, middle, end],
-                outputRange: [0, 1, 0],
-                extrapolate: 'clamp'
-            }) : 1;
+            let animatedValue;
 
             this._positions[index] = {
                 start: index * sizeRef,
                 end: index * sizeRef + sizeRef
             };
+
+            if (!this._shouldAnimateSlides(props)) {
+                animatedValue = new Animated.Value(1);
+            } else if (this._shouldUseCustomAnimation()) {
+                animatedValue = new Animated.Value(_index === this._activeItem ? 1 : 0);
+            } else {
+                let interpolator;
+
+                if (scrollInterpolator) {
+                    interpolator = scrollInterpolator(_index, props);
+                } else if (this._shouldUseStackLayout()) {
+                    interpolator = stackScrollInterpolator(_index, props);
+                } else if (this._shouldUseTinderLayout()) {
+                    interpolator = tinderScrollInterpolator(_index, props);
+                }
+
+                if (!interpolator || !interpolator.inputRange || !interpolator.outputRange) {
+                    interpolator = defaultScrollInterpolator(_index, props);
+                }
+
+                animatedValue = this._scrollPos.interpolate({
+                    ...interpolator,
+                    extrapolate: 'clamp'
+                });
+            }
 
             interpolators.push(animatedValue);
         });
@@ -492,10 +616,69 @@ export default class Carousel extends Component {
         this.setState({ interpolators });
     }
 
+    _getSlideAnimation (index, toValue) {
+        const { interpolators } = this.state;
+        const { activeAnimationType, activeAnimationOptions } = this.props;
+
+        const animatedValue = interpolators && interpolators[index];
+
+        if (!animatedValue && animatedValue !== 0) {
+            return null;
+        }
+
+        const animationCommonOptions = {
+            isInteraction: false,
+            useNativeDriver: true,
+            ...activeAnimationOptions,
+            toValue: toValue
+        };
+
+        return Animated.parallel([
+            Animated['timing'](
+                animatedValue,
+                { ...animationCommonOptions, easing: Easing.linear }
+            ),
+            Animated[activeAnimationType](
+                animatedValue,
+                { ...animationCommonOptions }
+            )
+        ]);
+    }
+
+    _playCustomSlideAnimation (current, next) {
+        const { interpolators } = this.state;
+        const itemsLength = this._getCustomDataLength();
+        const _currentIndex = this._getCustomIndex(current);
+        const _currentDataIndex = this._getDataIndex(_currentIndex);
+        const _nextIndex = this._getCustomIndex(next);
+        const _nextDataIndex = this._getDataIndex(_nextIndex);
+        let animations = [];
+
+        // Keep animations in sync when looping
+        if (this._enableLoop()) {
+            for (let i = 0; i < itemsLength; i++) {
+                if (this._getDataIndex(i) === _currentDataIndex && interpolators[i]) {
+                    animations.push(this._getSlideAnimation(i, 0));
+                } else if (this._getDataIndex(i) === _nextDataIndex && interpolators[i]) {
+                    animations.push(this._getSlideAnimation(i, 1));
+                }
+            }
+        } else {
+            if (interpolators[current]) {
+                animations.push(this._getSlideAnimation(current, 0));
+            }
+            if (interpolators[next]) {
+                animations.push(this._getSlideAnimation(next, 1));
+            }
+        }
+
+        Animated.parallel(animations, { stopTogether: false }).start();
+    }
+
     _hackActiveSlideAnimation (index, goTo, force = false) {
         const { data } = this.props;
 
-        if (IS_IOS || !this._flatlist || !this._positions[index] || (!force && this._enableLoop())) {
+        if (!this._mounted || !this._carouselRef || !this._positions[index] || (!force && this._enableLoop())) {
             return;
         }
 
@@ -508,26 +691,20 @@ export default class Carousel extends Component {
         const itemsLength = data && data.length;
         const direction = goTo || itemsLength === 1 ? 'start' : 'end';
 
-        this._flatlist && this._flatlist._listRef && this._flatlist.scrollToOffset({
-            offset: offset + (direction === 'start' ? -1 : 1),
-            animated: false
-        });
+        this._scrollTo(offset + (direction === 'start' ? -1 : 1), false);
 
         clearTimeout(this._hackSlideAnimationTimeout);
         this._hackSlideAnimationTimeout = setTimeout(() => {
-            // https://github.com/facebook/react-native/issues/10635
-            this._flatlist && this._flatlist._listRef && this._flatlist.scrollToOffset({
-                offset: offset,
-                animated: false
-            });
+            this._scrollTo(offset, false);
         }, 50); // works randomly when set to '0'
     }
 
     _lockScroll () {
+        const { lockScrollTimeoutDuration } = this.props;
         clearTimeout(this._lockScrollTimeout);
         this._lockScrollTimeout = setTimeout(() => {
             this._releaseScroll();
-        }, 1000);
+        }, lockScrollTimeoutDuration);
         this._setScrollEnabled(false);
     }
 
@@ -556,17 +733,49 @@ export default class Carousel extends Component {
         this._snapToItem(repositionTo, false, false, false, false);
     }
 
+    _scrollTo (offset, animated = true) {
+        const { vertical } = this.props;
+        const wrappedRef = this._getWrappedRef();
+
+        if (!this._mounted || !wrappedRef) {
+            return;
+        }
+
+        const specificOptions = this._needsScrollView() ? {
+            x: vertical ? 0 : offset,
+            y: vertical ? offset : 0
+        } : {
+            offset
+        };
+        const options = {
+            ...specificOptions,
+            animated
+        };
+
+        if (this._needsScrollView()) {
+            wrappedRef.scrollTo(options);
+        } else {
+            wrappedRef.scrollToOffset(options);
+        }
+    }
+
     _onScroll (event) {
-        const { enableMomentum, onScroll, callbackOffsetMargin } = this.props;
+        const { callbackOffsetMargin, enableMomentum, onScroll } = this.props;
 
         const scrollOffset = event ? this._getScrollOffset(event) : this._currentContentOffset;
         const nextActiveItem = this._getActiveItem(scrollOffset);
-        const scrollConditions = nextActiveItem === this._itemToSnapTo &&
+        const itemReached = nextActiveItem === this._itemToSnapTo;
+        const scrollConditions =
             scrollOffset >= this._scrollOffsetRef - callbackOffsetMargin &&
             scrollOffset <= this._scrollOffsetRef + callbackOffsetMargin;
 
         this._currentContentOffset = scrollOffset;
         this._onScrollTriggered = true;
+        this._lastScrollDate = Date.now();
+
+        if (this._activeItem !== nextActiveItem && this._shouldUseCustomAnimation()) {
+            this._playCustomSlideAnimation(this._activeItem, nextActiveItem);
+        }
 
         if (enableMomentum) {
             clearTimeout(this._snapNoMomentumTimeout);
@@ -575,18 +784,30 @@ export default class Carousel extends Component {
                 this._activeItem = nextActiveItem;
             }
 
-            if (scrollConditions && this._canFireCallback) {
-                this._onSnap(this._getDataIndex(nextActiveItem));
-            }
-        } else if (scrollConditions && this._activeItem !== nextActiveItem) {
-            this._activeItem = nextActiveItem;
+            if (itemReached) {
+                if (this._canFireBeforeCallback) {
+                    this._onBeforeSnap(this._getDataIndex(nextActiveItem));
+                }
 
-            if (this._canLockScroll()) {
-                this._releaseScroll();
+                if (scrollConditions && this._canFireCallback) {
+                    this._onSnap(this._getDataIndex(nextActiveItem));
+                }
+            }
+        } else if (this._activeItem !== nextActiveItem && itemReached) {
+            if (this._canFireBeforeCallback) {
+                this._onBeforeSnap(this._getDataIndex(nextActiveItem));
             }
 
-            if (this._canFireCallback) {
-                this._onSnap(this._getDataIndex(nextActiveItem));
+            if (scrollConditions) {
+                this._activeItem = nextActiveItem;
+
+                if (this._canLockScroll()) {
+                    this._releaseScroll();
+                }
+
+                if (this._canFireCallback) {
+                    this._onSnap(this._getDataIndex(nextActiveItem));
+                }
             }
         }
 
@@ -595,7 +816,7 @@ export default class Carousel extends Component {
             this._repositionScroll(nextActiveItem);
         }
 
-        if (onScroll && event) {
+        if (typeof onScroll === "function" && event) {
             onScroll(event);
         }
     }
@@ -611,9 +832,28 @@ export default class Carousel extends Component {
     }
 
     _onTouchStart () {
+        const { onTouchStart } = this.props
+
         // `onTouchStart` is fired even when `scrollEnabled` is set to `false`
         if (this._getScrollEnabled() !== false && this._autoplaying) {
-            this.stopAutoplay();
+            this.pauseAutoPlay();
+        }
+
+        if (onTouchStart) {
+            onTouchStart()
+        }
+    }
+
+    _onTouchEnd () {
+        const { onTouchEnd } = this.props
+
+        if (this._getScrollEnabled() !== false && this._autoplay && !this._autoplaying) {
+            // This event is buggy on Android, so a fallback is provided in _onScrollEnd()
+            this.startAutoplay();
+        }
+
+        if (onTouchEnd) {
+            onTouchEnd()
         }
     }
 
@@ -639,7 +879,7 @@ export default class Carousel extends Component {
     _onScrollEndDrag (event) {
         const { onScrollEndDrag } = this.props;
 
-        if (this._flatlist) {
+        if (this._carouselRef) {
             this._onScrollEnd && this._onScrollEnd();
         }
 
@@ -652,7 +892,7 @@ export default class Carousel extends Component {
     _onMomentumScrollEnd (event) {
         const { onMomentumScrollEnd } = this.props;
 
-        if (this._flatlist) {
+        if (this._carouselRef) {
             this._onScrollEnd && this._onScrollEnd();
         }
 
@@ -662,7 +902,7 @@ export default class Carousel extends Component {
     }
 
     _onScrollEnd (event) {
-        const { autoplay, enableSnap } = this.props;
+        const { autoplayDelay, enableSnap } = this.props;
 
         if (this._ignoreNextMomentum) {
             // iOS fix
@@ -677,15 +917,13 @@ export default class Carousel extends Component {
             this._snapScroll(this._scrollEndOffset - this._scrollStartOffset);
         }
 
-        if (autoplay) {
-            // Restart autoplay after a little while
-            // This could be done when releasing touch
-            // but the event is buggy on Android...
-            // https://github.com/facebook/react-native/issues/9439
+        // The touchEnd event is buggy on Android, so this will serve as a fallback whenever needed
+        // https://github.com/facebook/react-native/issues/9439
+        if (this._autoplay && !this._autoplaying) {
             clearTimeout(this._enableAutoplayTimeout);
             this._enableAutoplayTimeout = setTimeout(() => {
                 this.startAutoplay();
-            }, 300);
+            }, autoplayDelay + 50);
         }
     }
 
@@ -753,10 +991,11 @@ export default class Carousel extends Component {
     }
 
     _snapToItem (index, animated = true, fireCallback = true, initial = false, lockScroll = true) {
-        const { enableMomentum, onSnapToItem } = this.props;
+        const { enableMomentum, onSnapToItem, onBeforeSnapToItem } = this.props;
         const itemsLength = this._getCustomDataLength();
+        const wrappedRef = this._getWrappedRef();
 
-        if (!itemsLength || !this._flatlist || !this._flatlist._listRef) {
+        if (!itemsLength || !wrappedRef) {
             return;
         }
 
@@ -774,8 +1013,14 @@ export default class Carousel extends Component {
                 this._lockScroll();
             }
 
-            if (onSnapToItem && fireCallback) {
-                this._canFireCallback = true;
+            if (fireCallback) {
+                if (onBeforeSnapToItem) {
+                    this._canFireBeforeCallback = true;
+                }
+
+                if (onSnapToItem) {
+                    this._canFireCallback = true;
+                }
             }
         }
 
@@ -787,14 +1032,11 @@ export default class Carousel extends Component {
             return;
         }
 
-        this._flatlist && this._flatlist._listRef && this._flatlist.scrollToOffset({
-            offset: this._scrollOffsetRef,
-            animated
-        });
+        this._scrollTo(this._scrollOffsetRef, animated);
 
         if (enableMomentum) {
             // iOS fix, check the note in the constructor
-            if (IS_IOS && !initial) {
+            if (!initial) {
                 this._ignoreNextMomentum = true;
             }
 
@@ -815,10 +1057,21 @@ export default class Carousel extends Component {
         }
     }
 
+    _onBeforeSnap (index) {
+        const { onBeforeSnapToItem } = this.props;
+
+        if (!this._carouselRef) {
+            return;
+        }
+
+        this._canFireBeforeCallback = false;
+        onBeforeSnapToItem && onBeforeSnapToItem(index);
+    }
+
     _onSnap (index) {
         const { onSnapToItem } = this.props;
 
-        if (!this._flatlist) {
+        if (!this._carouselRef) {
             return;
         }
 
@@ -828,6 +1081,7 @@ export default class Carousel extends Component {
 
     startAutoplay () {
         const { autoplayInterval, autoplayDelay } = this.props;
+        this._autoplay = true;
 
         if (this._autoplaying) {
             return;
@@ -844,12 +1098,17 @@ export default class Carousel extends Component {
         }, autoplayDelay);
     }
 
-    stopAutoplay () {
+    pauseAutoPlay () {
         this._autoplaying = false;
-        clearInterval(this._autoplayInterval);
+        clearInterval(this._autoplayInterval);        
     }
 
-    snapToItem (index, animated = true) {
+    stopAutoplay () {
+        this._autoplay = false;
+        this.pauseAutoPlay();
+    }
+
+    snapToItem (index, animated = true, fireCallback = true) {
         if (!index || index < 0) {
             index = 0;
         }
@@ -860,10 +1119,10 @@ export default class Carousel extends Component {
             return;
         }
 
-        this._snapToItem(positionIndex, animated);
+        this._snapToItem(positionIndex, animated, fireCallback);
     }
 
-    snapToNext (animated = true) {
+    snapToNext (animated = true, fireCallback = true) {
         const itemsLength = this._getCustomDataLength();
 
         let newIndex = this._activeItem + 1;
@@ -873,10 +1132,10 @@ export default class Carousel extends Component {
             }
             newIndex = 0;
         }
-        this._snapToItem(newIndex, animated);
+        this._snapToItem(newIndex, animated, fireCallback);
     }
 
-    snapToPrev (animated = true) {
+    snapToPrev (animated = true, fireCallback = true) {
         const itemsLength = this._getCustomDataLength();
 
         let newIndex = this._activeItem - 1;
@@ -886,18 +1145,48 @@ export default class Carousel extends Component {
             }
             newIndex = itemsLength - 1;
         }
-        this._snapToItem(newIndex, animated);
+        this._snapToItem(newIndex, animated, fireCallback);
+    }
+
+    // https://github.com/facebook/react-native/issues/1831#issuecomment-231069668
+    triggerRenderingHack (offset) {
+        // Avoid messing with user scroll
+        if (Date.now() - this._lastScrollDate < 500) {
+            return;
+        }
+
+        const scrollPosition = this._currentContentOffset;
+        if (!scrollPosition && scrollPosition !== 0) {
+            return;
+        }
+
+        const scrollOffset = offset || (scrollPosition === 0 ? 1 : -1);
+        this._scrollTo(scrollPosition + scrollOffset, false);
+    }
+
+    _getSlideInterpolatedStyle (index, animatedValue) {
+        const { layoutCardOffset, slideInterpolatedStyle } = this.props;
+
+        if (slideInterpolatedStyle) {
+            return slideInterpolatedStyle(index, animatedValue, this.props);
+        } else if (this._shouldUseTinderLayout()) {
+            return tinderAnimatedStyles(index, animatedValue, this.props, layoutCardOffset);
+        } else if (this._shouldUseStackLayout()) {
+            return stackAnimatedStyles(index, animatedValue, this.props, layoutCardOffset);
+        } else if (this._shouldUseShiftLayout()) {
+            return shiftAnimatedStyles(index, animatedValue, this.props);
+        } else {
+            return defaultAnimatedStyles(index, animatedValue, this.props);
+        }
     }
 
     _renderItem ({ item, index }) {
         const { interpolators } = this.state;
         const {
-            inactiveSlideShift,
             hasParallaxImages,
-            inactiveSlideScale,
-            inactiveSlideOpacity,
             itemWidth,
             itemHeight,
+            keyExtractor,
             renderItem,
             sliderHeight,
             sliderWidth,
@@ -908,34 +1197,16 @@ export default class Carousel extends Component {
         const animatedValue = interpolators && interpolators[index];
 
         if (!animatedValue && animatedValue !== 0) {
-            return false;
+            return null;
         }
 
         const animate = this._shouldAnimateSlides();
         const Component = animate ? Animated.View : View;
-        const translateProp = vertical ? 'translateX' : 'translateY';
-
-        const animatedStyle = animate ? {
-            opacity: animatedValue.interpolate({
-                inputRange: [0, 1],
-                outputRange: [inactiveSlideOpacity, 1]
-            }),
-            transform: [{
-                scale: animatedValue.interpolate({
-                    inputRange: [0, 1],
-                    outputRange: [inactiveSlideScale, 1]
-                })
-            }, {
-                [translateProp]: animatedValue.interpolate({
-                    inputRange: [0, 1],
-                    outputRange: [inactiveSlideShift, 0]
-                })
-            }]
-        } : {};
+        const animatedStyle = animate ? this._getSlideInterpolatedStyle(index, animatedValue) : {};
 
         const parallaxProps = hasParallaxImages ? {
             scrollPosition: this._scrollPos,
-            carouselRef: this._flatlist,
+            carouselRef: this._carouselRef,
             vertical,
             sliderWidth,
             sliderHeight,
@@ -943,34 +1214,71 @@ export default class Carousel extends Component {
             itemHeight
         } : undefined;
 
+        const mainDimension = vertical ? { height: itemHeight } : { width: itemWidth };
+        const specificProps = this._needsScrollView() ? {
+            key: keyExtractor ? keyExtractor(item, index) : this._getKeyExtractor(item, index)
+        } : {};
+
         return (
-            <Component style={[slideStyle, animatedStyle]} pointerEvents={'box-none'}>
+            <Component style={[mainDimension, slideStyle, animatedStyle]} pointerEvents={'box-none'} {...specificProps}>
                 { renderItem({ item, index }, parallaxProps) }
             </Component>
         );
     }
 
-    render () {
+    _getComponentOverridableProps () {
+        const {
+            enableMomentum,
+            itemWidth,
+            itemHeight,
+            loopClonesPerSide,
+            sliderWidth,
+            sliderHeight,
+            vertical
+        } = this.props;
+
+        const visibleItems = Math.ceil(vertical ?
+            sliderHeight / itemHeight :
+            sliderWidth / itemWidth) + 1;
+        const initialNumPerSide = this._enableLoop() ? loopClonesPerSide : 2;
+        const initialNumToRender = visibleItems + (initialNumPerSide * 2);
+        const maxToRenderPerBatch = 1 + (initialNumToRender * 2);
+        const windowSize = maxToRenderPerBatch;
+
+        const specificProps = !this._needsScrollView() ? {
+            initialNumToRender: initialNumToRender,
+            maxToRenderPerBatch: maxToRenderPerBatch,
+            windowSize: windowSize
+            // updateCellsBatchingPeriod
+        } : {};
+
+        return {
+            decelerationRate: enableMomentum ? 0.9 : 'fast',
+            showsHorizontalScrollIndicator: false,
+            showsVerticalScrollIndicator: false,
+            overScrollMode: 'never',
+            automaticallyAdjustContentInsets: false,
+            directionalLockEnabled: true,
+            pinchGestureEnabled: false,
+            scrollsToTop: false,
+            removeClippedSubviews: !this._needsScrollView(),
+            inverted: this._needsRTLAdaptations(),
+            // renderToHardwareTextureAndroid: true,
+            ...specificProps
+        };
+    }
+
+    _getComponentStaticProps () {
         const { hideCarousel } = this.state;
         const {
             containerCustomStyle,
             contentContainerCustomStyle,
-            data,
-            enableMomentum,
-            itemWidth,
-            itemHeight,
             keyExtractor,
-            loopClonesPerSide,
-            renderItem,
             sliderWidth,
             sliderHeight,
             style,
             vertical
         } = this.props;
-
-        if (!data || !renderItem) {
-            return false;
-        }
 
         const containerStyle = [
             containerCustomStyle || style || {},
@@ -982,62 +1290,68 @@ export default class Carousel extends Component {
                 { width: sliderWidth, flexDirection: this._needsRTLAdaptations() ? 'row-reverse' : 'row' }
         ];
         const contentContainerStyle = [
-            contentContainerCustomStyle || {},
             vertical ? {
                 paddingTop: this._getContainerInnerMargin(),
                 paddingBottom: this._getContainerInnerMargin(true)
             } : {
                 paddingLeft: this._getContainerInnerMargin(),
                 paddingRight: this._getContainerInnerMargin(true)
-            }
+            },
+            contentContainerCustomStyle || {}
         ];
-        const visibleItems = Math.ceil(vertical ?
-            sliderHeight / itemHeight :
-            sliderWidth / itemWidth) + 1;
-        const initialNumPerSide = this._enableLoop() ? loopClonesPerSide : 2;
-        const initialNumToRender = visibleItems + (initialNumPerSide * 2);
-        const maxToRenderPerBatch = 1 + (initialNumToRender * 2);
-        const windowSize = maxToRenderPerBatch;
 
-        return (
-            <AnimatedFlatList
-              decelerationRate={enableMomentum ? 0.9 : 'fast'}
-              showsHorizontalScrollIndicator={false}
-              showsVerticalScrollIndicator={false}
-              overScrollMode={'never'}
-              automaticallyAdjustContentInsets={false}
-              directionalLockEnabled={true}
-              pinchGestureEnabled={false}
-              scrollsToTop={false}
-              initialNumToRender={initialNumToRender}
-              maxToRenderPerBatch={maxToRenderPerBatch}
-              windowSize={windowSize}
-              // updateCellsBatchingPeriod
-              // renderToHardwareTextureAndroid={true}
-              removeClippedSubviews={true}
-              inverted={this._needsRTLAdaptations()}
-              {...this.props}
-              ref={(c) => { if (c) { this._flatlist = c._component; } }}
-              data={this._getCustomData()}
-              renderItem={this._renderItem}
-              // extraData={this.state}
-              getItemLayout={undefined} // see #193
-              initialScrollIndex={undefined} // see #193
-              keyExtractor={keyExtractor || this._getKeyExtractor}
-              numColumns={1}
-              style={containerStyle}
-              contentContainerStyle={contentContainerStyle}
-              horizontal={!vertical}
-              scrollEventThrottle={1}
-              onScroll={this._onScrollHandler}
-              onScrollBeginDrag={this._onScrollBeginDrag}
-              onScrollEndDrag={this._onScrollEndDrag}
-              onMomentumScrollEnd={this._onMomentumScrollEnd}
-              onResponderRelease={this._onTouchRelease}
-              onStartShouldSetResponderCapture={this._onStartShouldSetResponderCapture}
-              onTouchStart={this._onTouchStart}
-              onLayout={this._onLayout}
-            />
+        const specificProps = !this._needsScrollView() ? {
+            // extraData: this.state,
+            renderItem: this._renderItem,
+            numColumns: 1,
+            keyExtractor: keyExtractor || this._getKeyExtractor
+        } : {};
+
+        return {
+            ref: c => this._carouselRef = c,
+            data: this._getCustomData(),
+            style: containerStyle,
+            contentContainerStyle: contentContainerStyle,
+            horizontal: !vertical,
+            scrollEventThrottle: 1,
+            onScroll: this._onScrollHandler,
+            onScrollBeginDrag: this._onScrollBeginDrag,
+            onScrollEndDrag: this._onScrollEndDrag,
+            onMomentumScrollEnd: this._onMomentumScrollEnd,
+            onResponderRelease: this._onTouchRelease,
+            onStartShouldSetResponderCapture: this._onStartShouldSetResponderCapture,
+            onTouchStart: this._onTouchStart,
+            onTouchEnd: this._onScrollEnd,
+            onLayout: this._onLayout,
+            ...specificProps
+        };
+    }
+
+    render () {
+        const { data, renderItem, useScrollView } = this.props;
+
+        if (!data || !renderItem) {
+            return null;
+        }
+
+        const props = {
+            ...this._getComponentOverridableProps(),
+            ...this.props,
+            ...this._getComponentStaticProps()
+        };
+
+        const ScrollViewComponent = typeof useScrollView === 'function' ? useScrollView : AnimatedScrollView
+
+        return this._needsScrollView() ? (
+            <ScrollViewComponent {...props}>
+                {
+                    this._getCustomData().map((item, index) => {
+                        return this._renderItem({ item, index });
+                    })
+                }
+            </ScrollViewComponent>
+        ) : (
+            <AnimatedFlatList {...props} />
         );
     }
 }