+++ CHANGELOG.md
@@ -1,3 +1,84 @@
+## v3.8.4
+
+* Remove TypeScript definitions since they were causing issues (thanks [@radko93](https://github.com/radko93))
+  * If you're using Typescript you should install the type definitions this way: `npm install --save @types/react-native-snap-carousel`
+
+## v3.8.3
+
+* Seems like it's time to remove the `getItemLayout` & `initialScrollIndex` override to properly display the first item when its index is a huge number (thanks [@rontalx](https://github.com/rontalx))
+  * :warning: Make sure to read how these props work together ([link#1](https://facebook.github.io/react-native/docs/flatlist#getitemlayout) & [link #2](https://facebook.github.io/react-native/docs/flatlist#initialscrollindex)), and to understand [why they were disabled in the first place](https://github.com/facebook/react-native/issues/15734#issuecomment-330616697)!
+* Fix `snapToItem` not working when `enableMomentum` is set to `true` on Android (thanks [@daaashleywch](https://github.com/daaashleywch))
+* Add basic TypeScript support (thanks [@facuacostag](https://github.com/facuacostag))
+* Stack layout's peaking items will remain opaque if `inactiveSlideOpacity` is set to `1`
+
+
+## v3.8.2
+
+* Fix autoplay stop after user interaction (thanks [@HelloCore](https://github.com/HelloCore))
+* Allow using a custom animated image component with `ParallaxImage` (thanks [@DanielMarkiel](https://github.com/DanielMarkiel))
+
+## v3.8.1
+
+* Migrate from deprecated `componentWillReceiveProps` to `componentDidUpdate` (thanks [@kiarashws](https://github.com/kiarashws))
+* Use `console.error` instead of `console.warn` for critical warnings (thanks [@bardiarastin](https://github.com/bardiarastin))
+* Update parallax doc (thanks [@bardiarastin](https://github.com/bardiarastin))
+
+## v3.8.0
+
+* Set [`removeClippedSubviews`](https://facebook.github.io/react-native/docs/scrollview#removeclippedsubviews) to `false` by default for 'tinder' and 'stack' layouts, or when `useScrollView` is set to `true`. This aims at preventing a bunch of rendering issues.
+* Make sure that autoplay is properly restarted after a `touchStart` event
+* Allow serialized animated event as `onScroll`. See [#439](https://github.com/archriss/react-native-snap-carousel/pull/439) for more info (thanks [@Jberivera](https://github.com/Jberivera))
+* Allow using a custom scroll component. See [#498](https://github.com/archriss/react-native-snap-carousel/pull/498) for more info (thanks [@martinezguillaume](https://github.com/martinezguillaume))
+* Prevent loop animation from being played when reaching the end of the dataset. See [#443](https://github.com/archriss/react-native-snap-carousel/pull/443) for more info (thanks [@suhanmoon](https://github.com/suhanmoon))
+* Fire the `onTouchStart` prop. See [#464](https://github.com/archriss/react-native-snap-carousel/pull/464) for more info (thanks [@sangle7](https://github.com/sangle7))
+* Add accessibilityLabel to `Pagination`. See [#438](https://github.com/archriss/react-native-snap-carousel/pull/438) for more info (thanks [@thymikee](https://github.com/thymikee))
+* Allow `contentContainerCustomStyle` to override default paddings. See [#482](https://github.com/archriss/react-native-snap-carousel/pull/482) for more info (thanks [@yamov](https://github.com/yamov))
+
+## v3.7.5
+* Fix issue with `scrollEnabled` introduced in version `3.7.3`... again! (thanks [@ifsnow](https://github.com/ifsnow))
+
+## v3.7.4
+* Fix issue with `scrollEnabled` introduced in version `3.7.3` (thanks [@JakeRawr](https://github.com/JakeRawr))
+
+## v3.7.3
+* Fix faulty animated value and make sure to always check for `data` before checking for `data.length`
+* Fix `scrollEnabled` override when it was initially set to `false` (thanks [@JakeRawr](https://github.com/JakeRawr))
+
+## v3.7.2
+* Fix `ParallaxImage` not being rendered (thanks [@louiszawadzki](https://github.com/louiszawadzki))
+
+## v3.7.1
+* Fix a potential crash in release mode (thanks [@hanpanpan200](https://github.com/hanpanpan200))
+* Do not round scroll offset's number in order to prevent potential issues with scroll repositioning
+
+## v3.7.0
+### New features and enhancements
+* Add a new callback method: [`onBeforeSnapToItem()`](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/PROPS_METHODS_AND_GETTERS.md#methods)
+* Add prop `lockScrollTimeoutDuration`
+* Add the ability to enable/disable callback's execution when snaping programmatically
+* Add the ability to change `scrollEnabled` to `true` from initially `false` (thanks [@tomauty](https://github.com/tomauty))
+### Bugfixes
+* Fix random errors when accessing wrapped component's reference
+* Fix errors triggered when calling `setState()` while the component has already been unmounted
+### Other
+* Support `keyExtractor` on `ScrollView` (thanks [@hadimhd](https://github.com/hadimhd))
+
+## v3.6.0
+* Add a `layout` prop to let users choose between 3 different carousel layouts (see [the documentation](https://github.com/archriss/react-native-snap-carousel#layouts-and-custom-interpolations))
+![react-native-snap-carousel default layout](https://i.imgur.com/e1WbZcu.gif)
+![react-native-snap-carousel stack layout](https://i.imgur.com/foMIGM2.gif)
+![react-native-snap-carousel tinder layout](https://i.imgur.com/R7OpEFs.gif)
+* Add the ability to define dynamic styles based on scroll position with props `scrollInterpolator` and `slideInterpolatedStyle`. This allows implementing custom animations and layouts (see [the dedicated documentation](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/CUSTOM_INTERPOLATIONS.md)).
+* Rename props `customAnimationType` and `customAnimationOptions` to `activeAnimationType` and `activeAnimationOptions`
+
+## v3.5.0
+* Add the ability to render either a `ScrollView` component or a `FlatList` one (default) ([see prop `useScrollView`](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/PROPS_METHODS_AND_GETTERS.md#behavior))
+* Add support for versions of React Native < `0.43` (see [this note](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/KNOWN_ISSUES.md#react-native-version))
+* Add support for custom animations ([see props `customAnimationType` and `customAnimationOptions`](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/PROPS_METHODS_AND_GETTERS.md#style-and-animation))
+* Add method [`triggerRenderingHack()`](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/PROPS_METHODS_AND_GETTERS.md#available-methods) to work around [a random `FlatList` bug](https://github.com/facebook/react-native/issues/1831) that keeps content hidden until the carousel is scrolled
+* Hack around `ScrollView`/`FlatList` image rendering issues on Android
+* Fix issue with tappable dots when loop is enabled
+
 ## v3.4.0
 * Fix `snapToItem` call that results in snapping to the wrong item when `loop` is enabled
 * Fix issue that, in some use cases, prevents every items but the initial ones to be rendered
@@ -24,7 +105,7 @@
 * Make tappable `PaginationDot` snaps to the right item when loop is enabled
 
 ## v3.3.0
-* Bring in the most wanted 'infinite loop' feature :tada: (see [the 'Loop' section](https://github.com/archriss/react-native-snap-carousel#loop) for more info about the new props `loop` and `loopClonesPerSide`)
+* Bring in the most wanted 'infinite loop' feature :tada: (see [the 'Loop' section](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/PROPS_METHODS_AND_GETTERS.md#loop) for more info about the new props `loop` and `loopClonesPerSide`)
 * Improve Android behavior when momentum is disabled
 * Guard against potential errors when component is unmounted
 * Add prop `lockScrollWhileSnapping` to improve behavior when momentum is disabled
@@ -61,14 +142,14 @@
 ### Breaking changes
 * Plugin is now built on top of `FlatList`, which allows for huge performance optimizations. From now on, items must be rendered using props `data` and `renderItem`.
 ### General
-* Add `ParallaxImage` component (see the specific documentation [here](https://github.com/archriss/react-native-snap-carousel/blob/master/src/parallaximage/README.md))
+* Add `ParallaxImage` component (see the specific documentation [here](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/PARALLAX_IMAGE.md))
 * Add prop `activeSlideAlignment`
 * Fix issue with autoplay when setting `scrollEnabled` to `false`
 * Prevent going back to the first item when overscrolling the last one
 * Prevent callback from being called at the wrong time in some specific scenarios
 
 ## v2.4.0
-* Add `Pagination` component (see the specific documentation [here](https://github.com/archriss/react-native-snap-carousel/blob/master/src/pagination/README.md))
+* Add `Pagination` component (see the specific documentation [here](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/PAGINATION.md))
 * Allow `firstItem` to be changed dynamically
 * Allow `0` value for `carouselHorizontalPadding` and `carouselVerticalPadding` (thanks [@bonbonez](https://github.com/bonbonez))
 * Keep the easing of slide's opacity animation linear
@@ -97,7 +178,7 @@
 * Implement vertical mode (prop `vertical`)
 * Make sure that current active item is properly updated when snapping
 * Prevent issues when 'sliderWidth' is smaller than viewport's width
-* Recalculate card positions on layout to handle rotation (thanks [@andrewpope](https://github.com/andrewpope)); make sure to read [this note](https://github.com/archriss/react-native-snap-carousel#handling-device-rotation)
+* Recalculate card positions on layout to handle rotation (thanks [@andrewpope](https://github.com/andrewpope)); make sure to read [this note](https://github.com/archriss/react-native-snap-carousel/blob/master/doc/TIPS_AND_TRICKS.md#handling-device-rotation)
 * Refresh card positions if slider and/or item's dimensions are updated (thanks [@hoangnm](https://github.com/hoangnm))
 * Add props `scrollEndDragThrottleValue` and `snapCallbackDebounceValue`
 * Expose `View`'s `onLayout` prop